+++ lib/MarkedInput/utils/TriggerFinder.tsx
@@ -1,35 +1,21 @@
 import {Options, Trigger} from "../types";
 import {escapeRegex} from "./index";
 import {wordRegex} from "../constants";
+import {Caret} from "./Caret";
 
 export class TriggerFinder {
     text: string;
     dividedText: { left: string; right: string }
 
     static find(options: Options) {
-        if (this.isContinue)
+        if (Caret.isSelectedPosition)
             return new TriggerFinder().find(options)
     }
 
-    static get isContinue() {
-        const selection = window.getSelection()
-        if (!selection) return false
-
-        return selection.anchorOffset === selection.focusOffset
-    }
-
     constructor() {
-        let caretPosition = this.getCaretPosition()
-        this.text = this.getCurrentPieceOfText()
+        let caretPosition = Caret.getCurrentPosition()
+        this.text = Caret.getCurrentPieceOfText()
         this.dividedText = this.getDividedTextBy(caretPosition)
-    }
-
-    getCaretPosition() {
-        return window.getSelection()?.anchorOffset ?? 0
-    }
-
-    getCurrentPieceOfText() {
-        return window.getSelection()?.anchorNode?.textContent ?? ""
     }
 
     getDividedTextBy(position: number) {
@@ -44,7 +30,6 @@ export class TriggerFinder {
                 source: match.annotation,
                 index: match.index,
                 piece: this.text,
-                style: this.getCaretAbsolutePosition(),
                 option
             }
         }
@@ -80,12 +65,5 @@ export class TriggerFinder {
     makeTriggerRegex(trigger: string): RegExp {
         const patten = escapeRegex(trigger) + '(\\w*)$'
         return new RegExp(patten)
-    }
-
-    getCaretAbsolutePosition() {
-        const rect = window.getSelection()?.getRangeAt(0).getBoundingClientRect()
-        if (rect)
-            return {left: rect.left, top: rect.top + 20}
-        return {left: 0, top: 0}
     }
 }
