--- tests/units/triggerToRegex.spec.ts
@@ -1,7 +1,0 @@
-import {triggerToRegex} from "rc-marked-input/MarkedInput/utils";
-
-describe.skip(`Utility: ${triggerToRegex.name}`, () => {
-    it('should ', () => {
-
-    });
-})
